move logging the separate module
