+++ batcore/bat_logging/__init__.py
@@ -1,0 +1,6 @@
+from .loggers import *
+
+__all__ = [
+    "Logger",
+    "tester_logging",
+]
